# API.yaml
openapi: 3.0.3
info:
  title: URL Shortner API
  version: "1.0.0"
  description: |
    Backend URL shortening service.

    - Method-level docs and this OpenAPI description were generated automatically by AI.
    - Base URL is configurable via `app.base.url` (defaults to `http://localhost:8080`).
servers:
  - url: http://localhost:8080
paths:
  /api/v1/shortenUrl:
    post:
      summary: Create or reuse a short URL
      description: Creates (or reuses) a short URL for a full URL with an optional alias.
      tags: [Shortening]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUrl'
      responses:
        "200":
          description: Short URL successfully created or reused.
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: Full short URL (including base, encoded id and optional alias).
        "400":
          $ref: '#/components/responses/BadRequestError'
  /api/v1/data:
    post:
      summary: Get short URL metadata
      description: Returns metadata for an existing short URL id (and optional alias).
      tags: [Shortening]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestShortUrl'
      responses:
        "200":
          description: Metadata for the requested short URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUrl'
        "400":
          $ref: '#/components/responses/BadRequestError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "410":
          $ref: '#/components/responses/GoneError'
...
components:
  schemas:
    RequestUrl:
      type: object
      properties:
        url:
          type: string
          format: uri
        alias:
          type: string
          nullable: true
      required:
        - url
    RequestShortUrl:
      type: object
      properties:
        url_id:
          type: string
          description: Base64-url encoded id of the stored short URL.
        alias:
          type: string
          nullable: true
      required:
        - url_id
    ResponseUrl:
      type: object
      properties:
        shortUrl:
          type: string
        fullUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        requestCount:
          type: integer
      required:
        - shortUrl
        - fullUrl
        - createdAt
        - expiresAt
        - requestCount
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
      required:
        - status
        - error
        - message
  responses:
    BadRequestError:
      description: 400 Bad Request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: 404 Not Found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GoneError:
      description: 410 Gone.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'